requirements.txt
Flask==2.3.6
Werkzeug==2.3.6
gunicorn==21.2.0
waitress==2.1.2
psycopg2-binary==2.9.7